requirements.txt
# ── LLM & Agent ──────────────────────────────────────────────
anthropic==0.40.0
langchain==0.3.7
langchain-anthropic==0.3.0
langchain-community==0.3.7
langgraph==0.2.50

# ── Vector store & embeddings ─────────────────────────────────
chromadb==0.5.20
sentence-transformers==3.3.1

# ── Document processing ───────────────────────────────────────
pypdf==5.1.0
python-docx==1.1.2

# ── API ───────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pydantic==2.10.2
pydantic-settings==2.6.1

# ── Web UI ────────────────────────────────────────────────────
streamlit==1.40.2
httpx==0.27.2

# ── Slack ─────────────────────────────────────────────────────
slack-bolt==1.21.2

# ── Auth & security ───────────────────────────────────────────
bcrypt==4.2.1
slowapi==0.1.9
PyJWT==2.9.0

# ── Database ──────────────────────────────────────────────────
psycopg2-binary>=2.9.10

# ── Evaluation ────────────────────────────────────────────────
ragas==0.2.15
datasets==3.2.0
openai==1.55.0

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2

# ── Testing ───────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
